+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 * Implement instances search query saving to cql file instead of csv. Refs UIDEXP-155.
 * Instance record (detailed view). Add copy number to item list. Refs UIIN-1251.
 * Instance record (detailed view). Add holdings copy number to holdings accordion. Refs UIIN-1254.
+* Use `<MultiSelectionFilter>` for item status filter. Refs UIIN-1224.
 
 ## [4.0.1](https://github.com/folio-org/ui-inventory/tree/v4.0.1) (2020-07-01)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v4.0.0...v4.0.1)